Removing unnecessary resources
