+++ app/src/main/kotlin/com/ledevs/bromate/app/contract/EntryContract.kt
@@ -5,10 +5,9 @@ import com.ledevs.bromate.app.ui.view.BaseView
 import com.ledevs.bromate.app.viewmodel.EntryViewModel
 import java.util.Date
 
-object MainContract {
+object EntryContract {
   interface Presenter : BasePresenter<View>
   interface View : BaseView {
-    fun getCurrentMonth(): Date
     fun showEntryList(entries: List<EntryViewModel>)
     fun showEntryLoadError()
   }
