+++ app/src/main/kotlin/com/ledevs/bromate/app/presenter/EntryPresenter.kt
@@ -1,22 +1,29 @@
 package com.ledevs.bromate.app.presenter
 
-import com.ledevs.bromate.app.contract.MainContract
+import com.ledevs.bromate.app.contract.EntryContract
 import com.ledevs.bromate.app.formatter.Formatter
 import com.ledevs.bromate.app.viewmodel.EntryViewModel
 import com.ledevs.bromate.data.repository.EntryRepository
 import io.reactivex.disposables.Disposables
 
-class MainPresenter(
+class EntryPresenter(
     private val formatter: Formatter,
     private val entryRepository: EntryRepository
-): MainContract.Presenter {
+): EntryContract.Presenter {
+
+  private lateinit var view: EntryContract.View
 
   private var subscription = Disposables.empty()
 
-  override fun attachView(view: MainContract.View) {
+  override fun attachView(view: EntryContract.View) {
+    this.view = view
+    requestOpenEntries()
+  }
+
+  private fun requestOpenEntries() {
     subscription.dispose()
 
-    subscription = entryRepository.listEntries(view.getCurrentMonth())
+    subscription = entryRepository.listOpenEntries()
         .map { EntryViewModel.createFrom(formatter, it) }
         .subscribe(
             {
