+++ app/src/main/kotlin/com/ledevs/bromate/app/ui/list/viewholder/EntryViewHolder.kt
@@ -1,4 +1,4 @@
-package com.ledevs.bromate.app.ui.adapters.viewholder
+package com.ledevs.bromate.app.ui.list.viewholder
 
 import android.databinding.ViewDataBinding
 import android.view.LayoutInflater
