+++ app/src/main/kotlin/com/ledevs/bromate/data/repository/EntryRepository.kt
@@ -5,5 +5,5 @@ import io.reactivex.Single
 import java.util.*
 
 interface EntryRepository {
-  fun listEntries(month: Date): Single<List<Entry>>
+  fun listOpenEntries(): Single<List<Entry>>
 }
