+++ app/src/main/kotlin/com/ledevs/bromate/di/module/ActivityBindModule.kt
@@ -1,19 +1,19 @@
 package com.ledevs.bromate.di.module
 
 import com.ledevs.bromate.di.ActivityInjectionBuilder
-import com.ledevs.bromate.di.subcomponent.MainActivityComponent
-import com.ledevs.bromate.app.ui.activities.MainActivity
+import com.ledevs.bromate.di.subcomponent.EntryActivityComponent
+import com.ledevs.bromate.app.ui.activities.EntryActivity
 import dagger.Binds
 import dagger.Module
 import dagger.multibindings.ClassKey
 import dagger.multibindings.IntoMap
 
 @Module(
-    subcomponents = arrayOf(MainActivityComponent::class)
+    subcomponents = arrayOf(EntryActivityComponent::class)
 )
 abstract class ActivityBindModule {
   @Binds
   @IntoMap
-  @ClassKey(MainActivity::class)
-  abstract fun bindsMainActivity(builder: MainActivityComponent.Builder): ActivityInjectionBuilder
+  @ClassKey(EntryActivity::class)
+  abstract fun bindsMainActivity(builder: EntryActivityComponent.Builder): ActivityInjectionBuilder
 }
