--- app/src/main/res/color/month_selector_tab_text_color.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-  <item android:state_checked="true" android:state_enabled="true" android:color="@android:color/white"/>
-  <item android:state_focused="true" android:state_enabled="true" android:color="@android:color/white"/>
-  <item android:state_selected="true" android:state_enabled="true" android:color="@android:color/white"/>
-  <item android:state_enabled="true" android:color="@color/windowBackground"/>
-</selector>
