--- app/src/main/res/drawable/border_white.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-  <item>
-    <shape>
-      <solid android:color="@android:color/white"/>
-      <stroke android:color="@color/windowBackground" android:width="0.5dp"/>
-    </shape>
-  </item>
-</selector>
