--- app/src/main/res/drawable/month_selector_gradient.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-  <gradient
-      android:startColor="#8185D2"
-      android:endColor="#6D90D5"
-      android:angle="-90"
-      android:type="linear"/>
-</shape>
