--- app/src/main/res/menu/bottom_navigation_menu.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-      xmlns:app="http://schemas.android.com/apk/res-auto">
-  <item android:id="@+id/menu_home"
-        android:title="Home"
-        android:icon="@drawable/ic_home_black_24dp"
-        app:showAsAction="always"/>
-  <item android:id="@+id/menu_profile"
-        android:title="Profile"
-        android:icon="@drawable/ic_person_pin_black_24dp"
-        app:showAsAction="always"/>
-</menu>
