+++ build.gradle
@@ -1,7 +1,7 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-  ext.kotlin_version = '1.1.1'
+  ext.kotlin_version = '1.1.2'
   ext.support_version = "25.3.1"
   ext.dagger_version = "2.10"
   ext.constraint_layout_version = "1.0.1"
